==> 01 CollectData.ipynb <==
## Collect Data
Collected data from the website "live.emx-timing.se".
Data will be collected from six urls = six different competitions in the same cup.
Since it seems to be a dynamically loaded page the data was saved manually into a CSVs

The CSV-files are imported into dataframes to sort out the columns to be able to read them in the right way.
Saving the dataframes into new CSV-files to ba able to use the data late and in next Jupyter Notebook.

Starting by importing

import os
import pandas as pd

Import data from CSV-files.

# Folder containing CSV files
folder_path = 'src/raw_data/'

# List all files in the folder ending with '.csv'
file_paths = [os.path.join(folder_path, file) for file in os.listdir(folder_path) if file.endswith('.csv')]

# Initialize a dictionary to store dataframes
dataframes = {}

# Loop through each file and read its data into a pandas DataFrame
for file_path in file_paths:
    # Extract file number from the filename (assuming the file name format is consistent)
    file_number = file_path.split('/')[-1].split('-')[0]
    
    # Read CSV into a dataframe and store it in the dictionary
    try:
        # Read the file skipping the first row and use it as header
        df = pd.read_csv(file_path, encoding='cp1252', header=None)
        # Use the second row as column names
        df.columns = df.iloc[0]
        # Drop the second row after using it as column names
        df = df.drop(0)
        dataframes[f'Dataframe_{file_number}'] = df
    except UnicodeDecodeError:
        print(f"Error reading {file_path}. Check encoding or file contents.")

Able to show the data from each csv.

dataframes['Dataframe_1'].head(-10)

dataframes['Dataframe_2'].head(3)

dataframes['Dataframe_3'].head(3)

dataframes['Dataframe_4'].head(3)

dataframes['Dataframe_5'].head(3)

dataframes['Dataframe_6'].head(3)

Start cleaning data.

print(dataframes['Dataframe_1'].columns)

First fix the 'Varvtider' column. The data contains several times for each lap. Split that into separate columns.

# Create a folder named lap_times_fixed if it doesn't exist
folder_name = 'src/lap_times_fixed'
if not os.path.exists(folder_name):
    os.makedirs(folder_name)

# Save each modified DataFrame to a CSV file
for key in dataframes.keys():
    df = dataframes[key]
    # Remove rows with NaN values in the 'Varv' column
    df.dropna(subset=['Varv'], inplace=True)
    
    # Convert 'Varv' column to integers
    df['Varv'] = df['Varv'].astype(int)
    
    # Perform operations on each dataframe here
    max_laps = df['Varv'].max()

    # Add columns Varv 1, Varv 2, Varv 3, etc., based on the maximum number of laps
    for i in range(1, max_laps + 1):
        df[f'Varv {i}'] = ''
    
    # Split 'Varvtider' column into separate columns based on the maximum number of laps
    split_varvtider = df['Varvtider'].str.split(expand=True)
    
    # Fill the new 'Varv' columns with the split lap times
    for i, col in enumerate(split_varvtider.columns, start=1):
        df[f'Varv {i}'] = split_varvtider[col]
    
    # Drop the original 'Varvtider' column
    df = df.drop(columns=['Varvtider','Märke/Anmälare'])
    
    # Save the modified DataFrame to a new CSV file
    file_path = os.path.join(folder_name, f'{key}.csv')
    df.to_csv(file_path, index=False)
    print(f"Saved {key} to {file_path}")

dataframes['Dataframe_1'].head(-10)
