==> 04 SeedForNewEvent.ipynb <==
## Seed riders for new event

This Notebook is to utilize the ranked riders data and load a list of participating riders in a new event.
The riders in the new event will get seeded if they have been participating in the competition from last year's Vintercupen.

import os
import pandas as pd

## Load previous ranked data

file_path = 'src/ranked/ranked.csv'

# Read the CSV file into a DataFrame
ranked_dataframe = pd.read_csv(file_path)

# Display the first few rows of the DataFrame
print(ranked_dataframe.head())

## load participants for an upcoming event.

file_path = 'src/unseeded/unseeded.csv'

# Read the CSV file into a DataFrame with the second row as column names
unseeded_dataframe = pd.read_csv(file_path, encoding='cp1252', header=0)

# Display the first few rows of the DataFrame
print(unseeded_dataframe.head(-10))

# Merge or lookup operation to find the AvgRank from ranked_dataframe for each row in unseeded_dataframe
seeded_df = unseeded_dataframe.merge(
    ranked_dataframe[['Namn', 'Klubb', 'AvgRank']],  # Selecting necessary columns from ranked_dataframe
    on=['Namn', 'Klubb'],  # Merging on 'Namn' and 'Klubb' columns
    how='left'  # Perform a left join to keep all rows from unseeded_dataframe
)

# Drop the 'Unnamed: 5' column if it exists
if 'Unnamed: 5' in seeded_df.columns:
    seeded_df.drop(columns=['Unnamed: 5'], inplace=True)

# Rename the 'AvgRank' column to 'Seeding'
seeded_df.rename(columns={'AvgRank': 'Seeding'}, inplace=True)

# Find the maximum AvgRank from ranked_dataframe
max_avg_rank = ranked_dataframe['AvgRank'].max()

# Fill missing Seeding values with the maximum AvgRank plus 1
seeded_df['Seeding'].fillna(max_avg_rank + 1, inplace=True)

# Sort by Seeding column
seeded_df = seeded_df.sort_values(by='Seeding')

seeded_df.head(-10)

seeded_df['Seeding'] = seeded_df.reset_index(drop=True).index + 1

seeded_df.head(100)

import datetime

folder_name = 'src/seeded'
if not os.path.exists(folder_name):
    os.makedirs(folder_name)
    
# Get the current date and time
current_time = datetime.datetime.now().strftime("%Y-%m-%d_%H%M")

# Define the filename with the current timestamp
file_name = f"src/seeded/{current_time}.csv"

# Save the DataFrame to a CSV file
try:
    seeded_df.to_csv(file_name, index=False, encoding='cp1252')  # Windows-1252 encoding
    print(f"Ranked data saved successfully to '{file_name}'")
except Exception as e:
    print(f"Error saving ranked data: {e}")
